move code from init to own file
