+++ rasa/nlu/extractors/crf_entity_extractor.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, List, Optional, Text, Tuple, Union
 from rasa.nlu.config import InvalidConfigError, RasaNLUModelConfig
 from rasa.nlu.extractors import EntityExtractor
 from rasa.nlu.model import Metadata
-from rasa.nlu.tokenizers import Token
+from rasa.nlu.tokenizers.tokenizer import Token
 from rasa.nlu.training_data import Message, TrainingData
 from rasa.constants import DOCS_BASE_URL
